fix: move EpisodeSidebarTabs to settings/controls to fix architecture check
Having both settings/modal/SettingsSidebar.tsx and shell/RightSidebar.tsx
import EpisodeSidebarTabs from shell/ created a settings<->shell
bidirectional dependency, which the multimodal domain architecture check
forbids. Moving it into settings/controls/ (the "reusable controls" domain)
keeps the dependency one-directional (shell already legitimately depends
on settings).
